refactor(event): 클래스명을 ChristmasDate에서 EventDate로 변경
