+++ src/main/java/christmas/services/date/DateReferee.java
@@ -1,6 +1,6 @@
 package christmas.services.date;
 
-import static christmas.enums.ChristmasDate.CHRISTMAS_DAY;
+import static christmas.enums.EventDate.CHRISTMAS_DAY;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
